Isabel's Final
This is a very rough first draft but I'm already having a little
trouble. Since I know that I will be using particles for my project, I
thought that I should first translate the previous particles exercise
from p5js to Processing. I'm having trouble with making the array and
the for loop working. I commented my questions in the sketch as well as
my plan for the next steps of the project.