final update 05
I tried to do a flow where I check if hero has run out of lives, in
which case the user would click start and everything would start over.
It’s in the sketch file, but I check if hero has run out of hearts in
the Hero class and return true/false to the main sketch. That’s
working. What’s not working is the starting over.