OrlandoDiaz_FinalProjectupdate
So I got a lot of stuff working! Yay! But the problem I’m having is
with the healthbar and health counters, I point out the line and
comment on what’s happening in the code and would appreciate help in
that area. And how would you go about making it so that the boss and
hero take turns attacking? It’s boggling my mind trying to comprehend
how to put that into code. Also, I added a little intro movement and
animation for my monster. Enjoy the music too! And speaking of the
music, how do I make it so that when you press “z” it doesn’t play the
song again when you press “z” again? That’s my current situation right
now.